requirements.txt
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==39.0.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-allauth==0.52.0
django-crispy-forms==2.0
djangorestframework==3.14.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.1
fasteners==0.14.1
Flask==2.2.2
future==0.18.2
h11==0.14.0
httplib2==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.2.0
netifaces==0.10.4
oauthlib==3.2.2
olefile==0.46
outcome==1.2.0
packaging==23.1
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==9.3.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.9.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
systemd-python==234
testresources==2.0.1
texttable==1.6.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.13
usb-creator==0.3.7
wadllib==1.3.3
webdriver-manager==3.8.6
websocket-client==0.53.0
Werkzeug==2.2.2
wsproto==1.2.0
xkit==0.0.0
zipp==3.15.0
